refactor: move newwallet to connectors
