+++ src/app/router/connectorRoutes.tsx
@@ -16,7 +16,7 @@ import ConnectLnbits from "../screens/connectors/ConnectLnbits";
 import ConnectGaloy, { galoyUrls } from "../screens/connectors/ConnectGaloy";
 import ConnectEclair from "../screens/connectors/ConnectEclair";
 import ConnectCitadel from "../screens/connectors/ConnectCitadel";
-import NewWallet from "../screens/Onboard/NewWallet";
+import NewWallet from "../screens/connectors/NewWallet";
 import ConnectUmbrel from "../screens/connectors/ConnectUmbrel";
 import ConnectStart9 from "../screens/connectors/ConnectStart9";
 import ConnectMyNode from "../screens/connectors/ConnectMyNode";
