+++ src/app/screens/connectors/NewWallet/index.tsx
